Move PR functions
These functions process PRs, so have been moved into the PRs module.
